Split build graph declared env tests
